Use tasks.named for integrationTestTaskAssigned
Previously Gretty used tasks."${taskName}" to access the farm integration
tasks. This caused the tasks to be initialized at configuration time. The
impact is very noticeable in projects that had a lot of projects. A real
world project running into this is Spring Security. Applying
the Gretty plugin in Spring Security's build results in the Gretty
Plugin's Collection callbacks taking about 3.6 seconds of unnecessary
configuration time.

To better isolate performance of the Gretty Plugin I created
https://github.com/rwinch/gretty-performance-configuration-callback In
this project applying the Gretty plugin results in the Gretty Plugin's
Collection callbacks taking about 5.5 seconds of configuration time.

A link to the scan is found below:
https://scans.gradle.com/s/nsn2j4x3qroa2

With these changes the Collection callbacks is reduced to 0.39 seconds.

A link to the improved scan is found below:
https://scans.gradle.com/s/g6crbfgfeqmfm

Issue gh-141