fix: Increase maxPartCount in Tomcat connector
In Tomcat 10.1.42 (which is used by gretty 4.1.7) the maxPartCount in
Tomcat was set to 10 which is too low for many applications.
This was then increased to 50 in Tomcat 10.1.43

The following exception is thrown if this max value is hit:
FileCountLimitExceededException

This commit change the value to unlimited, just like the maxPostSize
currently is.

---- 10.1.42 - Tomcat repository ----
Add:  Provide finer grained control of multi-part request processing
via two new attributes on the Connector element. maxPartCount limits
the total number of parts in a multi-part request and maxPartHeaderSize
limits the size of the headers provided with each part. Add support for
these new attributes to the ParameterLimitValve. (markt)

---- 10.1.43 - Tomcat repository ----
<bug>69710</bug>: Increase the default for <code>maxPartCount</code>
from <code>10</code> to <code>50</code>. Update the documentation to
provide more details on the memory requirements to support multi-part
uploads while avoiding a denial of service risk. (markt)